Move tests folder to same level as implementation
