+++ src/__tests__/.jshintrc
